Rename to PreBufferRequestStream...
