+++ src/SystemWebAdapters/src/Adapters/PreBufferRequestStreamAttribute.cs
@@ -4,7 +4,7 @@
 namespace System.Web.Adapters;
 
 [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
-public class BufferRequestStreamAttribute : Attribute, IBufferRequestStreamMetadata
+public class PreBufferRequestStreamAttribute : Attribute, IPreBufferRequestStreamMetadata
 {
     // Same limit as the default: https://source.dot.net/#Microsoft.AspNetCore.Http/Internal/BufferingHelper.cs,47b7015acb14f2a4
     private const int DefaultBufferThreshold = 1024 * 30;
